refactor(examples): extract convert module
